==> Tabular with Sigmoid/Part 1 - Prepare case base and models.ipynb <==
# Default of Credit Ratings Dataset
## 1: Clean Data, Save Case Base and Models

import pickle
import pandas as pd
import numpy as np

from keras.models import Sequential
from keras.layers import Dense, Activation

from sklearn.neighbors import KNeighborsClassifier
from sklearn.model_selection import cross_val_score, train_test_split
from sklearn.preprocessing import MinMaxScaler
from sklearn.metrics import accuracy_score, confusion_matrix

df = pd.read_excel("default of credit card clients.xls")

# Save target y and dummy the X matrix
y = df["credible"]
del df["credible"]
del df["ID"]

# One hot encode fully to get high dimensional space
df = pd.get_dummies(df, columns=["SEX"])
df = pd.get_dummies(df, columns=["EDUCATION"])
df = pd.get_dummies(df, columns=["MARRIAGE"])

# Scale the Training Data Matrix
scaler = MinMaxScaler(feature_range=(-1, 1), copy=True)
scaler.fit(df.values)
X = scaler.transform(df.values)
y = y.values

## Make Case Base: 

# Save modified df (to save column names etc.)
df.to_csv("processed_df.csv", index=False)

# Make case base
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

## Train a Keras MLP

model = Sequential()
model.add(Dense(len(X_train[0]), input_dim=len(X_train[0])))
model.add(Activation("relu"))
model.add(Dense(1))
model.add(Activation("sigmoid"))

model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])

model.fit(X_train, y_train, epochs=30, batch_size=8)

scores = model.evaluate(X_train, y_train)
print("Training Set:", "\n%s: %.2f%%" % (model.metrics_names[1], scores[1]*100))

scores = model.evaluate(X_test, y_test)
print("Test Set:", "\n%s: %.2f%%" % (model.metrics_names[1], scores[1]*100))

# Use brute for maximum reliability in experiments
knn_clf = KNeighborsClassifier(n_neighbors=1, algorithm="brute") 
knn_clf.fit(X_train, y_train)

knn_predictions_test = knn_clf.predict(X_test)
print("k-NN Accuracy Test:", accuracy_score(y_test, knn_predictions_test))

confusion_matrix(y_test, knn_predictions_test, labels=None, sample_weight=None)

confusion_matrix(y_test, model.predict_classes(X_test), labels=None, sample_weight=None)

pickle.dump(knn_clf, open('k-nn_model.sav', 'wb'))

model.save("NN.h5")

np.save("X_train", X_train)
np.save("X_test", X_test)
np.save("y_train", y_train)
np.save("y_test", y_test)
