==> Tabular with Sigmoid/Part 6 - Remove Outliers and Test Again.ipynb <==
# 6: Test Case Retrieval With Local Methods Using Deleted Data


import pandas as pd
import numpy as np
import pickle
import time

from sklearn.decomposition import PCA
from sklearn.metrics import accuracy_score, confusion_matrix
from sklearn.metrics.pairwise import euclidean_distances
from sklearn.preprocessing import MinMaxScaler
from sklearn.neighbors import KNeighborsClassifier

from collections import Counter

from copy import deepcopy

import lime
import lime.lime_tabular

from keras.models import Sequential, load_model
from keras.layers import Dense, Activation
from keras import backend as K

# My own algorithms
from Weighted_KNN_Classifier import *

%matplotlib inline

df = pd.read_csv("processed_df.csv")
feature_names = df.columns

## Load Data 

k = 1

model = load_model("NN.h5")

X_train = np.load("X_train.npy")
X_test = np.load("X_test.npy")
y_train = np.load("y_train.npy")
y_test = np.load("y_test.npy")

X_train_grad_LIME = np.load("X_train_grad_LIME.npy")
X_test_grad_LIME = np.load("X_test_grad_LIME.npy")

X_train_intgrad = np.load("X_train_intgrad.npy")
X_test_intgrad = np.load("X_test_intgrad.npy")
X_train_deeplift = np.load("X_train_deeplift.npy")
X_test_deeplift = np.load("X_test_deeplift.npy")
X_train_lrp = np.load("X_train_lrp.npy")
X_test_lrp = np.load("X_test_lrp.npy")

## Delete Outlier Data for Final Evaluation

del_idx = list()
for i in range(len(X_train)):
    if y_train[i] != model.predict_classes(np.array([X_train[i]]))[0][0]:
        del_idx.append(i)

# Modify shape for next loop
y_train = np.array([y_train]).T

to_modify = [X_train, y_train, X_train_grad_LIME,
             X_train_intgrad, X_train_deeplift, X_train_lrp]

modified = list()
for i in range(len(to_modify)):
    matrix = deepcopy(to_modify[i].tolist())
    for index in sorted(del_idx, reverse=True):
        del matrix[index]
    modified.append(matrix)

names = ['X_train_del', 'y_train_del', 'X_train_grad_LIME_del',
         'X_train_intgrad_del', 'X_train_deeplift_del', 'X_train_lrp_del']

for i in range(len(names)):
    matrix = np.array(modified[i])
    name = names[i]
    # for y_train
    if i == 1:
        matrix = matrix.T[0]
    np.save(name, matrix) 

X_train = np.load("X_train_del.npy")
X_test = np.load("X_test.npy")
y_train = np.load("y_train_del.npy")
y_test = np.load("y_test.npy")

X_train_grad_LIME = np.load("X_train_grad_LIME_del.npy")
X_train_intgrad = np.load("X_train_intgrad_del.npy")
X_train_deeplift = np.load("X_train_deeplift_del.npy")
X_train_lrp = np.load("X_train_lrp_del.npy")

## Test Global Weighting

connection_weights = np.load('connection_weights.npy').item()
garsons_weights = np.load('garsons_weights.npy').item()
sensitivity_weights = np.load('sensitivity_weights.npy').item()
perturb_weights = np.load('perturb_weights.npy').item()
global_weights = [perturb_weights, sensitivity_weights, garsons_weights, connection_weights]

final_weights = []
for i in range(len(global_weights)):
    weights = global_weights[i]
    weight_array = list()
    for j in range(len(feature_names)):
        if feature_names[j] in weights:
            weight_array.append(weights[feature_names[j]]) 
    final_weights.append(weight_array)

for i in range(len(final_weights)):
    for j in range(len(final_weights[i])):
        final_weights[i][j] = abs(final_weights[i][j])
final_weights = np.array(final_weights)

knn_clf = KNeighborsClassifier(n_neighbors=5, algorithm="brute") 
knn_clf.fit(X_train, y_train)

knn_pred = knn_clf.predict(X_test)
nn_pred = model.predict_classes(X_test)
print("NN Accuracy Test:", accuracy_score(y_test, nn_pred))
print("k-NN Accuracy Test:", accuracy_score(y_test, knn_pred))

print("=====================================")
print("Global Searches: k =", k)
print("=====================================")
for i in range(len(final_weights)):
    if i == 0:
        technique = "Perturbation"
    elif i == 1:
        technique = "Sensitivity"
    elif i == 2:
        technique = "Garson's Algorithm"
    elif i == 3:
        technique = "Connection Weights"
    clf = GlobalWeightedKNN()
    clf.fit(X_train, y_train, k=k, weights=final_weights[i])
    predictions = clf.predict(X_test)
    print(technique + ": ", accuracy_score(y_test, predictions))
    print(confusion_matrix(y_test, predictions, labels=None, sample_weight=None))
    print(" ")
    right = 0
    for i in range(len(nn_pred)):
        if predictions[i] == nn_pred[i]:
            right += 1
    print("Agreement:", right/len(nn_pred))
    print(" ")
    print("-------------------------------------")

## Test Local Weighting

nn_pred = model.predict_classes(X_test)

techniques = [["k-NN*", X_train, X_test],
              ["C-DeepLIFT", X_train_deeplift, X_test_deeplift],
              ["C-Integraded Gradients", X_train_intgrad, X_test_intgrad],
              ["C-LRP", X_train_lrp, X_test_lrp]]

for item in techniques:
    technique = item[0]
    train = item[1]
    test = item[2]
    print("==================================================")
    print(technique)
    print("==================================================")
    kNN = KNeighborsClassifier(n_neighbors=k, algorithm="brute") 
    kNN.fit(train, y_train)
    knn_predictions_test = kNN.predict(test)
    print("Accuracy:", accuracy_score(y_test, knn_predictions_test))
    print(confusion_matrix(y_test, knn_predictions_test, labels=None, sample_weight=None))
    print(" ")
    right = 0
    for i in range(len(nn_pred)):
        if knn_predictions_test[i] == nn_pred[i]:
            right += 1
    print("Agreement:", right/len(nn_pred))

## Weighted L2


weighted_knn = LocalWeightedKNN()
weighted_knn.fit(X_train, y_train, k=k)

techniques = [["Local Linear Model", X_test_grad_LIME[:, :len(feature_names)]]]

for item in techniques:
    technique = item[0]
    weights = item[1]
    print("==================================================")
    print(technique)
    print("==================================================")
    start = time.time()
    predictions = list()
    for i in range(len(X_test)):            
        prediction = weighted_knn.predict(X_test[i], abs(weights[i]))
        predictions.append(prediction)
    print("Accuracy: ", accuracy_score(y_test, predictions))
    print(confusion_matrix(y_test, predictions, labels=None, sample_weight=None))
    print("Time taken:", time.time() - start)
    print(" ")
    right = 0
    for i in range(len(nn_pred)):
        if predictions[i] == nn_pred[i]:
            right += 1
    print("Agreement:", right/len(nn_pred))
